Test creation of debugfs file.
We've added a static varable to `memory.c`. This should create a debugfs
file called `shoopus_ma_whoopus`. If we write a value to it, it'll store
the value incremented by one (unless we write 42, which will error).
Reading will work as normal. This will demonstrate our capability to
store state in `mm/memory.c`, and to print it to the debugfs.